+++ lib/src/components/_internal/Icons/ArrowDown.tsx
@@ -18,24 +18,3 @@ export const ArrowDown = () => {
     </svg>
   );
 };
-
-export const Clear = () => {
-  return (
-    <svg
-      xmlns="http://www.w3.org/2000/svg"
-      className="icon icon-tabler icon-tabler-circle-x"
-      width={18}
-      height={18}
-      viewBox="0 0 24 24"
-      strokeWidth="2"
-      stroke="currentColor"
-      fill="none"
-      strokeLinecap="round"
-      strokeLinejoin="round"
-    >
-      <path stroke="none" d="M0 0h24v24H0z" fill="none"></path>
-      <circle cx={12} cy={12} r={9}></circle>
-      <path d="M10 10l4 4m0 -4l-4 4"></path>
-    </svg>
-  );
-};
